REFACTOR: 컴포넌트 모달 수정
